Main EDA file updated
The EDA analysis where significantly updated in the comments but also in the coding part. The main rework focus on the autocorrelation analysis since having a dataset which is very large can reduce the significance of confidence bands. hence both ACF and PACF analysis where introduced along with stonger economic constraints for significance. Also, the section related to the Leverage effect and the part where the LOWESS function is called for estimating a curve fitting the bins that are calculated via a hybrid approach has been clarified via comments.